[#2] Reject absent filenames in PUT and DELETE
Absent filenames already resulted in status code 409 (conflict) as a
result of trying to overwrite or delete the www/data directory.  Now
they produce a more appropriate 405 (method not allowed).

+++ index.js
@@ -45,6 +45,7 @@ const getPath = async function (req, res) {
 };
 
 const putData = async function (req, res) {
+    if (!req.params._) return;
     let dataPath = "www/data/" + req.params._;
     await fs.ensureDir(path.dirname(dataPath));
     await fs.writeFile(dataPath, await buffer(req));
@@ -52,6 +53,7 @@ const putData = async function (req, res) {
 }
 
 const delData = async function (req, res) {
+    if (!req.params._) return;
     let dataPath = "www/data/" + req.params._;
     await fs.unlink(dataPath);
     try {
